mm: kasan: unified support for SLUB and SLAB allocators
This is a backport of
https://github.com/steelannelida/kasan/commit/170d2ccea1d4c16acea67b0956f21fa3169eaba9
to the chromeos-3.14 kernel.

Original commit description:

With this patch kasan can be compiled with both SLAB and SLUB
allocators,
using minimal dependencies on allocator internal structures and minimum
allocator-dependent code.

Dependency from SLUB_DEBUG is also removed. The metadata storage is made
more efficient, so the redzones aren't as large for small objects. The
redzone size is calculated based on the object size.

Stack storage isn't implemented here and is planned for a future patch

The gfp flags are added to kasan hooks for future patches as well.

Signed-off-by: Dmitry Chernenkov <dmitryc@google.com>