mm: kasan: stack depots
This is a backport of
https://github.com/steelannelida/kasan/commit/7c9b30f499dfd5f48b39fbbd0006c788bd72f72a
to the chromeos-3.14 kernel.

Original commit description:

This patch enables kasan to store the call stacks for allocations and
deallocations efficiently.

SLUB_DEBUG supports the SLAB_STORE_USER option which stores
alloc/dealloc
stacks, but it consumes over 300 bytes per object. This patch suggests a
better storage for stacks. It uses only 24 bytes per object (+ extra 24
bytes if the object still can be used after allocation) + shared
storage.

We store stacks in separately allocated pages, back-to-back. The stacks
are
never deallocated, so fragmentation is not a problem. Duplicate stacks
are
not stored, instead a hashtable is used to index the stack contents.
Only a
32-bit handle is stored in the object metadata. According to our tests,
on
the average the same stack is reused ~100 times.

Signed-off-by: Dmitry Chernenkov <dmitryc@google.com>