mm: kasan: quarantine
This is a backport of
https://github.com/steelannelida/kasan/commit/b9fa66aa2057eeb6a4f537b6edfb85e4961d06ea
to the chromeos-3.14 kernel.

Original commit description:

Quarantine isolates freed objects into a separate queue. The objects are
returned to the allocator later, which facilitates detecting
use-after-free

In this implementation, in the free fastpath the object is added to the
per-cpu quarantine queue first. Those then get dumped into the global
queue

The quarantine maximum size is a fraction of installed physical memory.

Also, in case SLUB allocator is used, a shrinker is registered for the
quarantine. For SLAB the size of quarantine is reduced.

Signed-off-by: Dmitry Chernenkov <dmitryc@google.com>